slightly refactor sourcemap utils
