+++ packages/core/utils/src/index.js
@@ -18,7 +18,6 @@ export {default as promisify} from './promisify';
 export {default as validateSchema} from './schema';
 export {default as TapStream} from './TapStream';
 export {default as urlJoin} from './urlJoin';
-export {default as loadSourceMapUrl} from './loadSourceMapUrl';
 export {default as relativeUrl} from './relativeUrl';
 export {default as createDependencyLocation} from './dependency-location';
 export {default as debounce} from './debounce';
@@ -42,3 +41,4 @@ export * from './relativeBundlePath';
 export * from './ansi-html';
 export * from './escape-html';
 export * from './escape-markdown';
+export * from './sourcemap';
